生产者消费者分开在不同类，用了阻塞队列，具体逻辑还没写
