--- Main/src/main/java/com/ljh/main/ScopeTask/Service/Producer1.java
@@ -1,4 +1,0 @@
-package com.ljh.main.ScopeTask.Service;
-
-public class Producer1 {
-}
