+++ Main/src/main/java/com/ljh/main/ScopeTask/pojo/Product.java
@@ -1,0 +1,10 @@
+package com.ljh.main.ScopeTask.pojo;
+
+import lombok.Data;
+
+//生产者生产的数据
+@Data
+public class Product {
+    private String textContent;
+
+}
